Fix mirroring
While reviewing the old code and fixing a small problem, I created a new
one. Noticed this while diffing data output between old and new versions
of this service. My bad. This should fix this.

The lookup for the previously translated URI did not work correctly, and
the memoisations of the translation was also wrong.