Change file format to N-Triples instead of Turtle
In other pipelines of this service, stream processing has recently been
implemented for files. This assumes that files are encoded in N-Triples
syntax, which can be streamed, but some pipelines produced Turtle
syntax, which uses prefixes and human readable syntax that is not
streamable. This hack makes all files N-Triples syntax as well.