Replace one step with three: subjects, objects, sameAs
Mirroring is: replacing the subjects, replacing the objects, and adding
`owl:sameAs` triples to link to the original URIs. Instead of doing this
in 1 step, do this in seperate steps. This is necessary because when
files are split, the splits may occur in random places, and the data
itself might be completely out of order. When translating subjects, we
can not already translate objects, since not all related objects are in
the chunk (from the file streaming process). This needs to wait until we
have a complete picture of what URI needs to be translated in which
other URI. Same for adding `owl:sameAs` triples.